requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
loguru>=0.7.0
pytest>=7.4.0
python-dateutil>=2.8.2
rich>=13.0.0
openai>=1.0.0
google-generativeai>=0.3.0